requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
Cartopy==0.24.1
certifi==2025.4.26
cffi==1.17.1
cfgrib==0.9.15.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
eccodes==2.41.0
findlibs==0.1.1
flexcache==0.3
flexparser==0.4
folium==0.19.5
fonttools==4.57.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
herbie-data==2025.5.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
MetPy==1.7.0
narwhals==1.37.1
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.7
plotly==6.0.1
pooch==1.8.2
protobuf==6.30.2
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
shapely==2.1.0
siphon==0.10.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.45.0
streamlit_folium==0.25.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
xarray==2025.4.0
xyzservices==2025.4.0
